Move 'architectureData' to common
