+++ libs/ui-lib/lib/cim/components/Agent/NoAgentsAlert.tsx
@@ -3,7 +3,7 @@ import { Alert, AlertVariant, List, ListComponent, ListItem } from '@patternfly/
 import { Link } from 'react-router-dom-v5-compat';
 import { useTranslation } from '../../../common/hooks/use-translation-wrapper';
 import { Trans } from 'react-i18next';
-import { CpuArchitecture } from '../../../common';
+import { architectureData, CpuArchitecture, SupportedCpuArchitecture } from '../../../common';
 
 const NoAgentsAlert = ({ cpuArchitecture }: { cpuArchitecture: CpuArchitecture }) => {
   const { t } = useTranslation();
@@ -11,19 +11,23 @@ const NoAgentsAlert = ({ cpuArchitecture }: { cpuArchitecture: CpuArchitecture }
     <Alert
       variant={AlertVariant.warning}
       title={t('ai:No available hosts with {{cpuArchitecture}} architecture were found', {
-        cpuArchitecture,
+        cpuArchitecture: architectureData[cpuArchitecture as SupportedCpuArchitecture].label,
       })}
       isInline
     >
       {t(
         'ai:The cluster can not be installed yet because there are no available hosts with {{cpuArchitecture}} architecture found. To continue:',
-        { cpuArchitecture },
+        { cpuArchitecture: architectureData[cpuArchitecture as SupportedCpuArchitecture].label },
       )}
       <br />
       <List component={ListComponent.ol}>
         <ListItem>
           <Trans t={t} cpuArchitecture={cpuArchitecture}>
-            ai:Add hosts with {{ cpuArchitecture }} architecture to an{' '}
+            ai:Add hosts with{' '}
+            {{
+              cpuArchitecture: architectureData[cpuArchitecture as SupportedCpuArchitecture].label,
+            }}{' '}
+            architecture to an{' '}
             <Link to="/multicloud/infrastructure/environments">infrastructure environment</Link>
           </Trans>
         </ListItem>
